+++ scaleMT/ScaleMTRouter/src/main/java/org/scalemt/router/logic/UserAdmissionControl.java
@@ -23,20 +23,20 @@ public class UserAdmissionControl {
         return instance;
     }
 
-    private int limitRegistered=10000000;
-    private int limitAnonymous=1000000;
+    private long limitRegistered=10000000;
+    private long limitAnonymous=1000000;
     private List<String> whiteListUsers;
     private List<String> whiteListIPs;
     public UserAdmissionControl() {
         try
         {
-            limitRegistered=Integer.parseInt(Util.readConfigurationProperty("user_limit_registered"));
+            limitRegistered=Long.parseLong(Util.readConfigurationProperty("user_limit_registered"));
         }
         catch(Exception e){}
 
         try
         {
-            limitAnonymous=Integer.parseInt(Util.readConfigurationProperty("user_limit_anonymous"));
+            limitAnonymous=Long.parseLong(Util.readConfigurationProperty("user_limit_anonymous"));
         }
         catch(Exception e){}
 
@@ -61,7 +61,7 @@ public class UserAdmissionControl {
 
     public boolean canTranslate(Requester rq)
     {
-        int cost =LoadBalancer.getInstance().getLoadPredictor().getRequestHistory().getCostUser(rq);
+        long cost =LoadBalancer.getInstance().getLoadPredictor().getRequestHistory().getCostUser(rq);
         boolean result=true;
 
         if(rq instanceof RegisteredRequester)
